publishing system

+++ mkenv.sh
@@ -1,24 +1,40 @@
 # Set up Python3 virtualenv
-python3 -m venv .env
-source .env/bin/activate
-curl https://raw.githubusercontent.com/pypa/pip/master/contrib/get-pip.py | python
-deactivate
-source .env/bin/activate
-pip install -r requirements.txt
+if [ ! -d .env ] ; then
+	python3 -m venv .env
+	source .env/bin/activate
+	curl https://raw.githubusercontent.com/pypa/pip/master/contrib/get-pip.py | python
+	deactivate
+	source .env/bin/activate
+	pip install -r requirements.txt
+else
+	source .env/bin/activate
+fi
 
 # Set up Haskell virtualenv
-hsenv
+if [ ! -d .hsenv ] ; then
+	hsenv
+fi
 source .hsenv/bin/activate
 
 # Install redo
-git clone https://github.com/Zankoku-Okuno/redo.git
-cabal install redo/
-redo/mk_redo_links.sh
-rm -rf redo/
+if ! `which redo`; then
+	git clone https://github.com/Zankoku-Okuno/redo.git
+	cabal install redo/
+	redo/mk_redo_links.sh
+	rm -rf redo/
+fi
 
 # Set up redo
-redo-init
+if [ ! -d .redo ] ; then
+	redo-init
+fi
 
-# Install my mako command
-cp mako .env/bin/
-chmod +x .env/bin/mako
+# Install my commands and libraries
+cp -v mako .env/bin/; chmod +x .env/bin/mako
+cp -v setconfig.py .env/bin/setconfig; chmod +x .env/bin/setconfig
+cp -v setconfig.py .env/lib/python3.*/site-packages/
+cp -v publish.py .env/bin/publish; chmod +x .env/bin/publish
+
+# Shut down the virtualenvs
+deactivate_hsenv
+deactivate
